+++ src/main/java/com/example/demo/entities/OrderDetail.java
@@ -34,7 +34,7 @@ public class OrderDetail implements Serializable {
 	@Id
 	@EmbeddedId
 	private OrderDetailKey id = new OrderDetailKey();
-	private int weight;
+	private Double weight;
 	private Double price;
 
 	@MapsId("productId")
